requirements.txt
mcp
fastmcp
requests
httpx
openai
python-dotenv
uvicorn
